refactor: 공통 사용 함수와 interface 정의 외부 파일로 분리
