+++ src/types/Plan.ts
@@ -1,0 +1,7 @@
+export interface TaskInfo {
+  name: string;
+  id: string;
+  status: number;
+  labels: string[];
+  deadline: string;
+}
